NLP/babycoach_proj/requirements.txt
fastapi
uvicorn[standard]
langgraph
openai
python-dotenv
pytest
numpy
rank-bm25
sentence-transformers
torch

